clients/http-python/pyproject.toml
[project]
name = "doc-translater-http-api-client"
version = "0.1.0"
description = "Generated Python client for the DocTranslater HTTP API"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "AGPL-3.0" }
dependencies = [
    "attrs>=22.2.0",
    "httpx>=0.23.0,<0.29.0",
    "python-dateutil>=2.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["doc_translater_http_api_client"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "I", "UP"]
